refactor(audit): factor signAuditBundle + BUNDLE_CANONICALIZATION_SPEC into service; add snapshot executor param
Spec PR5 §6.2: single source of signing truth. Move the checkpoint sign +
events-content digest + best-effort checkpoint persistence out of the
/audit/bundle route into services/audit.signAuditBundle so the upcoming
/evidence route reuses it verbatim (no behavior change to /audit/bundle).

Also add an optional AuditReadExecutor (snapshot tx) param to verifyAuditChain
and readAuditBundleData so PR5 case assembly can run chain-verify + bundle read
inside ONE coherent snapshot (§4.1/§4.3, KC06). Defaults to getDb() — existing
callers unchanged.

Route now imports signAuditBundle + BUNDLE_CANONICALIZATION_SPEC from the
service; dropped the now-unused local hexToBytes, checkpoint imports, and
auditCheckpoints import. api typechecks clean.

Refs: PR5-CASE-EVIDENCE-SPEC §6.2 [sol-orch]
